Sync naromero77 fork with upstream master

+++ cime/scripts/lib/e3sm_cime_mgmt.py
@@ -1,6 +1,7 @@
 from CIME.utils import run_cmd, run_cmd_no_fail, expect, get_timestamp, CIMEError
 
 import getpass, logging, os
+import stat as osstat
 
 # Constants
 ESMCI_REMOTE_NAME = "esmci_remote_for_split"
@@ -12,14 +13,15 @@ MERGE_TAG_PREFIX = "to-acme-"
 def setup():
 ###############################################################################
     run_cmd_no_fail("git config merge.renameLimit 999999")
-    run_cmd_no_fail("git checkout master && git pull", verbose=True)
+    run_cmd_no_fail("git checkout master && git pull && git submodule sync && git submodule update --init", verbose=True)
 
     remotes = run_cmd_no_fail("git remote")
     if ESMCI_REMOTE_NAME not in remotes:
         run_cmd_no_fail("git remote add {} {}".format(ESMCI_REMOTE_NAME, ESMCI_URL), verbose=True)
 
-    run_cmd_no_fail("git fetch --prune {}".format(ESMCI_REMOTE_NAME), verbose=True)
-    run_cmd_no_fail("git fetch --prune {} --tags".format(ESMCI_REMOTE_NAME), verbose=True)
+    for origin in ["origin", ESMCI_REMOTE_NAME]:
+        run_cmd_no_fail("git fetch --prune {}".format(origin), verbose=True)
+        run_cmd_no_fail("git fetch --prune {} --tags".format(origin), verbose=True)
 
     run_cmd_no_fail("git clean -fd", verbose=True)
 
@@ -86,23 +88,28 @@ def do_subtree_split(new_split_tag, merge_tag):
 def touches_file(start_range, end_range, filepath, title, skip=None):
 ###############################################################################
     skip_str = "--grep={} --invert-grep".format(skip) if skip is not None else ""
-    result = run_cmd_no_fail("git log {} {}..{} -- {}".format(skip_str, start_range, end_range, filepath))
+    result = run_cmd_no_fail("git log --first-parent {} {}..{} -- {}".format(skip_str, start_range, end_range, filepath))
 
     if result:
-        logging.debug("  touched by {} within range {}..{} by commits\n{}".format(title, start_range, end_range, result))
+        logging.info("    Touched by {} within range {}..{} by commits\n{}".format(title, start_range, end_range, result))
 
     return result != ""
 
 ###############################################################################
 def reset_file(version, srcpath, dstpath):
 ###############################################################################
-    os.remove(dstpath)
+    is_exe = os.access(dstpath, os.X_OK)
+    if os.path.exists(dstpath):
+        os.remove(dstpath)
     try:
         run_cmd_no_fail("git show {}:{} > {}".format(version, srcpath, dstpath))
     except CIMEError:
         # If the above failes, then the file was deleted
         run_cmd_no_fail("git rm -f {}".format(dstpath))
     else:
+        if is_exe:
+            os.chmod(dstpath, os.stat(dstpath).st_mode | osstat.S_IXUSR | osstat.S_IXGRP | osstat.S_IXOTH)
+
         run_cmd_no_fail("git add {}".format(dstpath))
 
 ###############################################################################
@@ -111,7 +118,7 @@ def get_last_instance_of(branch_name, head):
     return run_cmd_no_fail("git log --first-parent {} --grep='{}' -1 --oneline".format(head, branch_name)).split()[0]
 
 ###############################################################################
-def handle_easy_conflict(dst_filepath, is_merge):
+def handle_easy_conflict(dst_filepath, is_merge, src_head, dst_head, last_src_operation, last_dst_operation, src_branch_prefix):
 ###############################################################################
     """
     src = repo we are coming from
@@ -125,51 +132,22 @@ def handle_easy_conflict(dst_filepath, is_merge):
     src = cime
     dst = e3sm
     """
-    src_tag_prefix = MERGE_TAG_PREFIX if is_merge else SPLIT_TAG_PREFIX
-    dst_tag_prefix = SPLIT_TAG_PREFIX if is_merge else MERGE_TAG_PREFIX
-
-    # For split, src_tag_prefix = acme-split-, dst_tag_prefix = to-acme-
-
-    src_branch_prefix = "branch-for-{}".format(dst_tag_prefix) # NOTE: opposite of tag
-    dst_branch_prefix = "branch-for-{}".format(src_tag_prefix) # NOTE: opposite of tag
-
-    # For split, dst_branch_prefix = branch-for-acme-split
-
-    # we can't use ORIG_HEAD for src_head for splits
-    # because ORIG_HEAD has been re-written by subtree, making the tag
-    # we want to use useless in a range operation
-    src_head = "MERGE_HEAD" if is_merge else "origin/master"
-    dst_head = "ORIG_HEAD"  if is_merge else "MERGE_HEAD"
-
-    # For split, src_head = origin/master, dst_head = MERGE_HEAD
-
-    src_filepath = dst_filepath.replace("cime/", "", 1) if is_merge else os.path.join("cime", dst_filepath)
+    logging.info("  Examining file {} ...".format(dst_filepath))
 
     # For split, src_filepath = cime/xxx/yyy, dst_filepath = xxx/yyy
-
-    # There's no tag for last dst operation
-    try:
-        last_dst_operation = get_last_instance_of(dst_branch_prefix, dst_head)
-    except Exception as e:
-        logging.warning("Could not get most recent merge for branch {}, {}".format(dst_branch_prefix, e))
-        return False
-
-    # Use tag for last src operation
-    last_src_operation = get_tag(src_tag_prefix, expected_num=2)[0]
-
-    logging.info("Examining file {} ...".format(dst_filepath))
+    src_filepath = dst_filepath.replace("cime/", "", 1) if is_merge else os.path.join("cime", dst_filepath)
 
     if not touches_file(last_dst_operation, dst_head, dst_filepath, "dst"):
-        logging.info("  File '{}' appears to have had no recent dst mods, setting to src".format(dst_filepath))
+        logging.info("    File '{}' appears to have had no recent dst mods, setting to src".format(dst_filepath))
         reset_file(src_head, src_filepath, dst_filepath)
         return True
     # We don't want to pick up the last dst->src operation as a src modification of this file
     elif not touches_file(last_src_operation, src_head, src_filepath, "src", skip=src_branch_prefix):
-        logging.info("  File '{}' appears to have had no recent src mods, setting to dst".format(src_filepath))
+        logging.info("    File '{}' appears to have had no recent src mods, setting to dst".format(src_filepath))
         reset_file(dst_head, dst_filepath, dst_filepath)
         return True
     else:
-        logging.info("  File '{}' appears to have real conflicts".format(dst_filepath))
+        logging.info("    File '{}' appears to have real conflicts".format(dst_filepath))
         return False
 
 ###############################################################################
@@ -179,9 +157,38 @@ def handle_easy_conflicts(is_merge):
     if not conflicting_files:
         expect(False, "Merge appears to have failed for reasons other than merge conflicts")
 
+    # For split, src_tag_prefix = acme-split-, dst_tag_prefix = to-acme-
+    src_tag_prefix = MERGE_TAG_PREFIX if is_merge else SPLIT_TAG_PREFIX
+    dst_tag_prefix = SPLIT_TAG_PREFIX if is_merge else MERGE_TAG_PREFIX
+
+    # For split, src_branch_prefix = branch-for-to-acme-, dst_branch_prefix = branch-for-acme-split-
+    src_branch_prefix = "branch-for-{}".format(dst_tag_prefix) # NOTE: opposite of tag
+    dst_branch_prefix = "branch-for-{}".format(src_tag_prefix) # NOTE: opposite of tag
+
+    # we can't use ORIG_HEAD for src_head for splits
+    # because ORIG_HEAD has been re-written by subtree, making the tag
+    # we want to use useless in a range operation
+    #
+    # For split, src_head = origin/master, dst_head = MERGE_HEAD
+    src_head = "MERGE_HEAD" if is_merge else "origin/master"
+    dst_head = "ORIG_HEAD"  if is_merge else "MERGE_HEAD"
+
+    # There's no tag for last dst operation
+    try:
+        last_dst_operation = get_last_instance_of(dst_branch_prefix, dst_head)
+        logging.info("  Detecting last dst operation was: {}".format(last_dst_operation))
+    except Exception as e:
+        logging.warning("  Could not get most recent merge for branch {}, {}. ABORTING auto conflict resolution".\
+                            format(dst_branch_prefix, e))
+        return conflicting_files
+
+    # Use tag for last src operation
+    last_src_operation = get_tag(src_tag_prefix, expected_num=2)[0]
+    logging.info("  Detecting last src operation was: {}".format(last_src_operation))
+
     rv = []
     for conflicting_file in conflicting_files:
-        able_to_handle = handle_easy_conflict(conflicting_file, is_merge)
+        able_to_handle = handle_easy_conflict(conflicting_file, is_merge, src_head, dst_head, last_src_operation, last_dst_operation, src_branch_prefix)
         if not able_to_handle:
             rv.append(conflicting_file)
 
@@ -190,13 +197,20 @@ def handle_easy_conflicts(is_merge):
 ###############################################################################
 def handle_conflicts(is_merge=False, auto_conf=False):
 ###############################################################################
+    logging.info("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
     logging.info("There are conflicts, analyzing...")
+    logging.info("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+
     remaining_conflicts = handle_easy_conflicts(is_merge) if auto_conf else True
     if remaining_conflicts:
         expect(False, "There are merge conflicts. Please fix, commit, and re-run this tool with --resume")
     else:
         logging.info("All conflicts were automatically resovled, continuing")
         run_cmd_no_fail("git commit --no-edit")
+
+    logging.info("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+    logging.info("Automatic conflict resolution complete")
+    logging.info("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
 
 ###############################################################################
 def do_subtree_pull(squash=False, auto_conf=False):
